chore(github_core): sync copilot-cli lib mirror after api split
Mirror of .claude/lib/github_core/ into src/copilot-cli/lib/ produced by the
lib build step (build/scripts/build_all.py). Keeps the Copilot install set in
sync with the canonical package.

Refs #1910
